backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
decorator==5.1.1
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
executing==1.2.0
ipython
jedi
Jinja2
jsonpickle
Markdown
MarkupSafe
matplotlib-inline
parso
pexpect
pickleshare
prompt-toolkit
psycopg2
ptyprocess
pure-eval
Pygments
python-dotenv
pytz
six
sqlparse
stack-data
traitlets
wcwidth
